separated linear and task id-er
